A potential bug fix in mean_climate_driver.py
The driver contains the following code block:
    result_dict["Variable"] = dict()
    result_dict["Variable"]["id"] = varname
    if level is not None:
        result_dict["Variable"]["level"] = level * 100  # hPa to Pa

that attempts to convert the specified pressure level from hPa to Pa.
However, this code block conflicted with the following code block in
the associated libarary model "load_and_regrid.py":

if "plev" in list(ds.coords.keys()):
            if ds.plev.units == "Pa":
                level = level * 100  # hPa to Pa
            try:
                ds = ds.sel(plev=level)

The plev in reanalysis such as ERA5 has a unit of Pa by default. The
above two code blocks will lead to wrong level
and failure of the data selection.