Propose a change for floating number format for mean climate metrics output
Current compute_metrics.py used either ".5f" or ".3f" to form the
floating numbers for the output from the mean climate driver. As the
magnitude of the model variable can range from 1e5 (e.g pressure ) to
1e-3 (e.g. humidity), used the ".5f" and ".3f" can sometimes truncate
the small values to zero. It seems that using  a standard scientific
notation can be benificial in case the user processed the data without
converting the unit data ahead. In this way, it also provide
possibilities for users to convert the units within the diagnostic
datasets.